Move uploadReducer into separate file
